Fix client certificates
