+++ cloud-gateway-service/src/main/java/org/zowe/apiml/cloudgatewayservice/x509/ClientCertFilterFactory.java
@@ -8,7 +8,7 @@
  * Copyright Contributors to the Zowe Project.
  */
 
-package org.zowe.apiml.cloudgatewayservice.filters;
+package org.zowe.apiml.cloudgatewayservice.x509;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cloud.gateway.filter.GatewayFilter;
@@ -18,8 +18,6 @@ import org.springframework.stereotype.Service;
 import org.zowe.apiml.constants.ApimlConstants;
 
 import java.security.cert.CertificateEncodingException;
-import java.security.cert.X509Certificate;
-import java.util.Base64;
 
 import static org.zowe.apiml.constants.ApimlConstants.HTTP_CLIENT_USE_CLIENT_CERTIFICATE;
 
@@ -50,19 +48,16 @@ public class ClientCertFilterFactory extends AbstractGatewayFilterFactory<Client
         return ((exchange, chain) -> {
             ServerHttpRequest request = exchange.getRequest().mutate().headers(headers -> {
                 headers.remove(CLIENT_CERT_HEADER);
-                if (exchange.getRequest().getSslInfo() != null) {
-                    X509Certificate[] certificates = exchange.getRequest().getSslInfo().getPeerCertificates();
-                    if (certificates != null && certificates.length > 0) {
-                        try {
-                            final String encodedCert = Base64.getEncoder().encodeToString(certificates[0].getEncoded());
-                            exchange.getAttributes().put(HTTP_CLIENT_USE_CLIENT_CERTIFICATE, Boolean.TRUE);
-                            headers.add(CLIENT_CERT_HEADER, encodedCert);
-                            log.debug("Incoming client certificate has been added to the {} header.", CLIENT_CERT_HEADER);
-                        } catch (CertificateEncodingException e) {
-                            log.debug("Failed to encode the incoming client certificate. Error message: {}", e.getMessage());
-                            headers.add(ApimlConstants.AUTH_FAIL_HEADER, "Invalid client certificate in request. Error message: " + e.getMessage());
-                        }
+                try {
+                    final String encodedCert = X509Util.getEncodedClientCertificate(exchange.getRequest().getSslInfo());
+                    if (encodedCert != null) {
+                        exchange.getAttributes().put(HTTP_CLIENT_USE_CLIENT_CERTIFICATE, Boolean.TRUE);
+                        headers.add(CLIENT_CERT_HEADER, encodedCert);
+                        log.debug("Incoming client certificate has been added to the {} header.", CLIENT_CERT_HEADER);
                     }
+                } catch (CertificateEncodingException e) {
+                    log.debug("Failed to encode the incoming client certificate. Error message: {}", e.getMessage());
+                    headers.add(ApimlConstants.AUTH_FAIL_HEADER, "Invalid client certificate in request. Error message: " + e.getMessage());
                 }
             }).build();
             return chain.filter(exchange.mutate().request(request).build());
