+++ cloud-gateway-service/src/test/java/org/zowe/apiml/cloudgatewayservice/x509/ClientCertFilterFactoryTest.java
@@ -8,7 +8,7 @@
  * Copyright Contributors to the Zowe Project.
  */
 
-package org.zowe.apiml.cloudgatewayservice.filters;
+package org.zowe.apiml.cloudgatewayservice.x509;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
